requirements.txt
requests
pandas
numpy
python-dotenv
transformers
torch
streamlit
rapidfuzz
sentence-transformers
